create new methods for last session and tests for it
