+++ cmake_modules/FindSQLite3.cmake
