+++ src/main.cpp
@@ -1,6 +1,15 @@
 #include <iostream>
-
+#include "../include/physics/SQLite.h"
 int main() {
+	SQLiteDataBase db;
+	
+	//db.setUser("player5", "1");
+	//cout << db.getAuthorizeUser("player4", "1") << '\n';
+	//cout << db.setLastSession();
+	//cout << db.getAuthorizeUser("player2", "1") << '\n';
+	//cout << db.setLastSession();
+	cout << db.getLastSession();
+	cout << db.getUserNickname();
 	std::cout << "Physics Module ready!" << std::endl;
 	return 0;
 }
